backend/requirements.txt
google-cloud-bigquery==2.34.4
Flask==2.2.0
flask-cors==3.0.10
google-auth==2.6.2
requests==2.27.1
requests-toolbelt==0.9.1
